Do not return results if they are only partial
This fixes a bug where people were kicked out of the org during the
nightly sync when one of the threads were throttled until it crashed.

This lead to the phone number not being filled in, and the diff then
treated it as a member without phone number => can not have accessy.